Optionally filter jobs by node
Jobs have a three element column called `attempted_by`, where the first
field is the name of the node that attempted to run them. This makes
jobs filterable by clicking on the node list, just like queues and
states.

There isn't an index backing this filter, so it may be slow in a view
with hundreds of thousands of jobs (i.e. completed).

Closes #16