Prevent errors on mount during app reboot
For many app's supervision trees it is common for the phoenix endpoint
to start before oban does during boot. During an application restart
this causes errors as the dashboard attempts to mount and the oban
config isn't available yet.

This guards against that situation and gracefully pauses mounting until
oban is ready, or a 5 second timeout has ellapsed, whichever is first.

Closes sorentwo/oban#415