Use 'english' tsvectors and omit tags by default
Using the 'english' config for tsvector operations makes them
idempotent and indexable. The array operation for tsvector doesn't
support a config, and tags are no longer searched by default to prevent
an extraneous search for a column that's not frequently used.

In addition, the searching guide now includes index suggestions for
performance.

Closes sorentwo/oban#614