Improved json path queries and append on click
Kills two birds with one stone. Previously, clicking a suggested value
injected it into the input but didn't submit it as a filter. That was
unexpected, and led to an odd "deletion" of previous values when typing
multiple filters at once. Clicking on a suggested value immediately
submits it as a filter now.

Auto submission exposed an issue with json path completion, i.e. path
suggestions lacked a final `:` or `.` to indicate if it was a full
qualifier or part of a path.

Closes #78